+++ packages/checkbox-group/checkbox-group.ts
@@ -93,6 +93,9 @@ export class WarpCheckboxGroup extends FormControlMixin(LitElement) {
 	@property({ type: Boolean, reflect: true })
 	invalid = false;
 
+	@state()
+	private _hasHelpTextSlot = false;
+
 	// Track whether the user has interacted with the group.
 	#hasInteracted = false;
 
@@ -144,36 +147,50 @@ export class WarpCheckboxGroup extends FormControlMixin(LitElement) {
 		}
 	`;
 
+	helpTextSlotChange() {
+		const el = this.renderRoot.querySelector(
+			"slot[name=help-text]",
+		) as HTMLSlotElement;
+		const helpText = el.assignedElements();
+		if (helpText.length) this._hasHelpTextSlot = true;
+	}
+
 	render() {
 		const hasSelection = this.#getCheckedCount() > 0;
 		const requiredInvalid = this.required && !hasSelection;
 		const showRequiredError = requiredInvalid && this.#hasInteracted;
 		const isInvalid = this.invalid || showRequiredError;
 		const helpText = isInvalid ? this.#getRequiredMessage() : this.helpText;
-		const helpId = helpText ? "checkbox-group__help" : undefined;
+		const hasHelpText = Boolean(helpText || this._hasHelpTextSlot);
+		const helpId = hasHelpText ? "checkbox-group__help" : undefined;
 		const labelId = this.label ? "checkbox-group__label" : undefined;
 		const ariaInvalid = isInvalid ? "true" : undefined;
 
 		return html`
 			<div class="wrapper" tabindex="${this._internalTabIndex}">
-				${this.label
-					? html`
-							<div class="label" id="${labelId}">
-								<span>${this.label}</span>
-								${this.optional && !this.required
-									? html`
-											<span class="optional">
-												${i18n._({
-													id: "checkbox-group.label.optional",
-													message: "Optional",
-													comment: "Shown behind label when marked as optional",
-												})}
-											</span>
-										`
-									: nothing}
-							</div>
-						`
-					: nothing}
+				${
+					this.label
+						? html`
+								<div class="label" id="${ifDefined(labelId)}">
+									<span>${this.label}</span>
+									${
+										this.optional && !this.required
+											? html`
+													<span class="optional">
+														${i18n._({
+															id: "checkbox-group.label.optional",
+															message: "Optional",
+															comment:
+																"Shown behind label when marked as optional",
+														})}
+													</span>
+												`
+											: nothing
+									}
+								</div>
+							`
+						: nothing
+				}
 				<div
 					class="checkbox-group"
 					role="group"
@@ -183,14 +200,17 @@ export class WarpCheckboxGroup extends FormControlMixin(LitElement) {
 				>
 					<slot></slot>
 				</div>
-				${helpText
-					? html`<div
-							class="${isInvalid ? "help-text error" : "help-text"}"
-							id="${helpId}"
-						>
-							${helpText}
-						</div>`
-					: nothing}
+				<div
+					?hidden=${!hasHelpText}
+					class="${isInvalid ? "help-text error" : "help-text"}"
+					id="${ifDefined(helpId)}"
+				>
+					${helpText}
+					<slot
+						@slotchange="${this.helpTextSlotChange}"
+						name="help-text"
+					></slot>
+				</div>
 			</div>
 		`;
 	}