+++ packages/slider-thumb/slider-thumb.ts
@@ -688,11 +688,13 @@ class WarpSliderThumb extends FormControlMixin(LitElement) {
 			(this.suffix ?? "") !== "";
 		return html`
 			<div class="w-slider-thumb">
-				${!("anchorName" in document.documentElement.style)
-					? html`<div class="polyfill-range">
-							<div class="polyfill-active-range"></div>
-						</div>`
-					: nothing}
+				${
+					!("anchorName" in document.documentElement.style)
+						? html`<div class="polyfill-range">
+								<div class="polyfill-active-range"></div>
+							</div>`
+						: nothing
+				}
 				<input
 					id="range"
 					aria-label="${this.ariaLabel}"
@@ -718,24 +720,30 @@ class WarpSliderThumb extends FormControlMixin(LitElement) {
 					@keydown="${this.#onRangeSliderKeyDown}"
 				/>
 
-				${this.slot === "from" || !this.slot
-					? // avoid including these labels twice, for screen readers
-						html` <span class="sr-only">
-							${i18n.t({
-								id: "slider.label.from",
-								message: "From",
-							}) +
-							" " +
-							(this.labelFormatter ? this.labelFormatter("from") : this.min) +
-							", " +
-							i18n.t({
-								id: "slider.label.to",
-								message: "To",
-							}) +
-							" " +
-							(this.labelFormatter ? this.labelFormatter("to") : this.max)}
-						</span>`
-					: nothing}
+				${
+					this.slot === "from" || !this.slot
+						? // avoid including these labels twice, for screen readers
+							html` <span class="sr-only">
+								${
+									i18n.t({
+										id: "slider.label.from",
+										message: "From",
+									}) +
+									" " +
+									(this.labelFormatter
+										? this.labelFormatter("from")
+										: this.min) +
+									", " +
+									i18n.t({
+										id: "slider.label.to",
+										message: "To",
+									}) +
+									" " +
+									(this.labelFormatter ? this.labelFormatter("to") : this.max)
+								}
+							</span>`
+						: nothing
+				}
 
 				<span aria-hidden="true" class="w-slider-thumb__from-marker">
 					${this.labelFormatter ? this.labelFormatter("from") : this.min}
@@ -755,10 +763,12 @@ class WarpSliderThumb extends FormControlMixin(LitElement) {
 					tabindex="${this._hiddenTextfield ? -1 : nothing}"
 					placeholder="${this.placeholder}"
 					.value="${this.textFieldDisplayValue}"
-					.formatter=${this.valueFormatter && !showPlaceholder
-						? (value: string) =>
-								this.valueFormatter!(value, this.slot as SliderSlot)
-						: nothing}
+					.formatter=${
+						this.valueFormatter && !showPlaceholder
+							? (value: string) =>
+									this.valueFormatter!(value, this.slot as SliderSlot)
+							: nothing
+					}
 					min="${this.openEnded ? nothing : this.min}"
 					max="${this.openEnded ? nothing : this.max}"
 					step="${ifDefined(this.step)}"
@@ -767,12 +777,14 @@ class WarpSliderThumb extends FormControlMixin(LitElement) {
 					@keydown="${this.#onInputFieldKeyDown}"
 					?disabled="${this.disabled}"
 				>
-					${(this.suffix ?? "")
-						? html`<w-affix
-								slot="suffix"
-								label="${this.suffix ?? ""}"
-							></w-affix>`
-						: nothing}
+					${
+						(this.suffix ?? "")
+							? html`<w-affix
+									slot="suffix"
+									label="${this.suffix ?? ""}"
+								></w-affix>`
+							: nothing
+					}
 				</w-textfield>
 				<w-attention
 					tooltip