+++ packages/tab/react.ts
@@ -19,9 +19,12 @@ const BaseTab = createComponent({
 
 type BaseTabProps = React.ComponentPropsWithoutRef<typeof BaseTab>;
 
-type TabProps = Omit<BaseTabProps, "aria-selected"> & {
+type TabProps = Omit<BaseTabProps, "aria-selected" | "tabindex"> & {
+	htmlFor?: string;
 	/** An ID is required to avoid hydration issues */
 	id: string;
+	/** A tabIndex is required to avoid hydration issues. Set to 0 for the active tab (usually the first tab) and -1 for all other tabs. */
+	tabIndex: 0 | -1;
 	ariaSelected?: "true" | "false";
 };
 
@@ -31,9 +34,10 @@ type TabProps = Omit<BaseTabProps, "aria-selected"> & {
  * [Warp component reference](https://warp-ds.github.io/docs/components/tabs/frameworks/elements)
  */
 export const Tab = React.forwardRef<WarpTab, TabProps>(
-	({ ariaSelected, ...props }, ref) =>
+	({ ariaSelected, htmlFor, ...props }, ref) =>
 		React.createElement(BaseTab, {
 			slot: "tabs",
+			for: htmlFor,
 			...props,
 			...(ariaSelected !== undefined ? { "aria-selected": ariaSelected } : {}),
 			ref,