+++ packages/textfield/react.ts
@@ -28,9 +28,9 @@ type BaseTextFieldProps = React.ComponentPropsWithoutRef<typeof BaseTextField>;
 
 type TextFieldProps = Omit<
 	BaseTextFieldProps,
-	"help-text" | "min-length" | "max-length" | "read-only"
+	"help-text" | "helpText" | "min-length" | "max-length" | "read-only"
 > & {
-	helpText?: string;
+	helpText?: string | React.ReactElement;
 	minLength?: number;
 	maxLength?: number;
 	readOnly?: boolean;
@@ -42,20 +42,30 @@ type TextFieldProps = Omit<
  * [Warp component reference](https://warp-ds.github.io/docs/components/text-area/frameworks/elements)
  */
 export const TextField = React.forwardRef<WarpTextField, TextFieldProps>(
-	({ helpText, minLength, maxLength, readOnly, ...props }, ref) =>
-		React.createElement(BaseTextField, {
-			...props,
-			...(helpText !== undefined ? { "help-text": helpText } : {}),
-			...(minLength !== undefined ? { "min-length": minLength } : {}),
-			...(maxLength !== undefined ? { "max-length": maxLength } : {}),
-			...(readOnly ? { "read-only": true } : {}),
-			ref,
-		} as React.ComponentProps<typeof BaseTextField> & {
-			"help-text"?: string;
-			"min-length"?: number;
-			"max-length"?: number;
-			"read-only"?: boolean;
-		}),
+	({ helpText, minLength, maxLength, readOnly, ...props }, ref) => {
+		return React.createElement(
+			BaseTextField,
+			{
+				...props,
+				...(typeof helpText === "string" ? { "help-text": helpText } : {}),
+				...(minLength !== undefined ? { "min-length": minLength } : {}),
+				...(maxLength !== undefined ? { "max-length": maxLength } : {}),
+				...(readOnly ? { "read-only": true } : {}),
+				ref,
+			},
+			[
+				props.children,
+				// support taking in JSX in helpText and placing it in the correct slot on behalf of users
+				typeof helpText !== "undefined" && typeof helpText !== "string"
+					? React.createElement(
+							"div",
+							{ slot: "help-text" },
+							helpText as React.ReactElement,
+						)
+					: null,
+			],
+		) as React.ReactElement<TextFieldProps>;
+	},
 );
 
 TextField.displayName = "TextField";